requirements.txt
--index-url https://artefact.skao.int/repository/pypi-all/simple
python-dotenv
pyyaml
ska-sdp-config
ska-ser-logging
kubernetes
